+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * [x] Rename `key_pressed` to `key_press`
 * [x] Remove the name parameter from `mouse_click`.
 * [x] Remove `TF.hovered`. Left mouse click & drag is now handled by the `pan_zoom` object.
-* [x] Add left mouse interaction support to `image` and `frame`. New arguments `drag_tag` and `down_tag`.
+* [x] Add left mouse interaction support to `image` and `frame`. New arguments `drag_tag` and `down_tag`, and `hover_tag`.
   - `pan_zoom.is_hovered` is no longer present.
   - `image` event handling was substantially reworked. Starting the drag outside image window no longer triggers panning.
 * [ ] Make `polylines` handle a color list